% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplyr.R
\name{multiplyr}
\alias{multiplyr}
\title{Multiply two numbers}
\usage{
multiplyr(a, b)
}
\arguments{
\item{a}{A number}

\item{b}{A number}
}
\value{
The product of \code{a} and \code{b}
}
\description{
Multiply two numbers
}
\examples{
multiplyr(2, 5)
multiplyr(4, 8)
}
