% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nullfunc.R
\name{nullfunc}
\alias{nullfunc}
\title{A not-useful-at-all function}
\usage{
nullfunc()
}
\description{
A not-useful-at-all function
}
\details{
This is a function which returns NULL. It is there just for the sake of having a function
in the package.
}

