% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Poppack-package.R
\docType{data}
\name{pop2016}
\alias{pop2016}
\title{Population of Mashhad in 2016.}
\format{
A data frame with 188 rows and 11 variables:
\describe{
\item{OBJECTID}{Id of each block as a factor}
\item{population}{Number of whole people living in a block}
\item{Family}{Number of families living in a block}
\item{Owners}{Number of people who has their own property in a block}
\item{Tenants}{Number of people who rent a property in a block}
\item{Residential Units}{Whole number of residencial suits in a block}
\item{Apartments}{Number of suits in apartments in a block}
\item{Houses}{Number of village houses in a block}
\item{Cottage or shed}{Number of sheds in a block}
\item{Shape_Length}{Surrounding a block in meter scale}
\item{Shape_Area}{Surface area of a block in meter scale}
}
}
\source{
(https://WWW.amar.org.ir)
}
\usage{
data(pop2016)
}
\description{
A dataset containing information about 188 world countries, including
population estimates, GDP per capita, energy consumption estimates,
and CO_2 emissions.
}
\examples{
 data(pop2016)
 GDPPC <- pop2016$GDPPC
}
\keyword{datasets}
