% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Poppack-package.R
\docType{package}
\name{Poppack-package}
\alias{Poppack}
\alias{Poppack-package}
\title{Poppack: Population of Mashhad}
\description{
This package has some census data about Mashhad.

This package is an exercise in package creation using
R studio. The package includes a sample function and
a sample dataset with their respective documentation.
}
\references{
\url{https://github.com/NiloofarNL/Poppack}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/NiloofarNL/Poppack}
}

}
\author{
\strong{Maintainer}: Niloofar Nalaee \email{nalaeen@mcmaster.ca} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})


Niloofar Nalaee, School of Earth, Environment and Society, McMaster University \email{nalaeen@mcmaster.ca}
}
\keyword{internal}
